Re-write:
- How bots run
- Rendering and State handling
- Misc (other refactors along the way)

Rationale:

The code was getting hard to manage, with years of fixes and changes.
This takes code from a lot of experiments made outside of shoebot, and integrates the ideas here.
This should make it easier to maintain and extend.

[TODO]
How bots run:

This code was complex, and inside a Nodebotgrand-parent class, "Bot".

This will be replaced with a runner.   This idea came from Nodebox, but was then extended in external notebooks and brought in here.

[In progress]
Graphics
........

How graphics work is re-written, this should help maintainability as well as make it easier to use Shoebot with various graphics tech.

Rendering:
The rendering code was hard to understand, and heavily integrated with Cairo and Gtk.
Graphics has been seperated out, by using an intermediate format (handled by the state code), which can be passed to processed by different graphics backends.

The aims here are to make things easier to maintain, as well as opening the door to new graphics backends (it may be possible to port these from related projects).

Graphics State:

Graphics state is now defined as just the information needed by a backend to draw some Shoebot object (Beziercurve / Image) etc.
These are stored in using dataclasses along with some helpers.

This simplifies the work the interface to the backend graphics, while leaving the behaviour (and sometimes complex Nodebox based APIs) in the Grobs themselves.

A declarative interface [TODO - example here] has been added to add state to Graphics objects, while there is some complexity in the implementation it should be balanced by the removable of much adhoc state handling code.

Misc (other refactors along the way)
....................................

Various constants, now all moved inside enums.

Tasks
-----

[TODO]
- [x] Run simple bot without crashing
- [ ] Run unit tests
- [ ] Output to file
- [ ] Output to a window (ideally more than one type
- [ ] Variables (GUI)
- [ ] Variables (Sockets)
- [ ] Jupyter